feat: ✨ setup base registry contract
