+++ packages/global-nft-registry/src/browser.ts
@@ -1,22 +1,21 @@
-import {SmartContract as CoreSmartContract} from './SmartContract';
-import {SmartContractOptions as BaseSmartContractOptions} from './options/SmartContractOptions';
+import {GlobalNftRegistry as CoreSmartContract} from './global-nft-registry';
 import {InjectedAccount} from '@polkadot/extension-inject/types';
 import {web3Enable, web3FromAddress} from '@polkadot/extension-dapp';
 import {AddressOrPair, Signer} from '@polkadot/api/types';
-import {isKeyringPair} from '@polkadot/api/util';
 
-export {DEVNET, TESTNET, MAINNET} from './options/SmartContractOptions';
+import {GlobalNftRegistryOptions as BaseOptions} from './options';
+export {DEVNET, TESTNET, MAINNET} from './options';
 
-export type SmartContractOptions = BaseSmartContractOptions & {
+export type GlobalNftRegistryOptions = BaseOptions & {
     extensionOriginName?: string;
 };
 
 const isInjectedAccount = (input: any): input is InjectedAccount => typeof input.address === 'string';
 
-export class SmartContract extends CoreSmartContract {
+export class GlobalNftRegistry extends CoreSmartContract {
     static async buildAndConnect(
         accountOrSecretPhrase: InjectedAccount | AddressOrPair,
-        options?: SmartContractOptions,
+        options?: GlobalNftRegistryOptions,
     ) {
         if (isInjectedAccount(accountOrSecretPhrase)) {
             await web3Enable(options?.extensionOriginName || 'ddc');
