+++ packages/global-nft-registry/src/smart-contract-base.ts
