+++ packages/global-nft-registry/src/types/contract-types.ts
@@ -1,8 +1,11 @@
+import {Codec} from '@polkadot/types/types';
+
 export type AccountId = string;
 export type TokenId = bigint;
 export type ContractAddress = string;
 export type ChainId = bigint;
 export type Balance = bigint;
+export type Offset = bigint;
 
 export enum Role {
     TrustedService,
@@ -19,3 +22,12 @@ export interface TokenOwner {
     tokenId: TokenId;
     owner: AccountId;
 }
+
+/**
+ * Converts Codec instance to primitive js representation
+ *
+ * TODO: figure out how to auto-map u64 to bigint instead of hex string
+ */
+export const toJs = (codec: Codec) => {
+    return codec.toJSON();
+};
