+++ packages/global-nft-registry/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './contract-types';
 export * from './event-types';
+export * from './cere-types';
